+++ src/main/java/fr/curie/miclearning/apposeplugin/Sam3Parameters.java
@@ -10,8 +10,8 @@ public class Sam3Parameters {
     public static final double DEFAULT_BOX_IOU_THRESHOLD = 0.15;
     public static final int DEFAULT_REMOVE_AFTER_MISSING = 5;
     public static final int DEFAULT_MAX_SIDE_LENGTH_DETECT = 1008;
-    public static final int DEFAULT_MAX_SIDE_LENGTH_TRACK = 504; // Reduce this to increase speed at the cost of mask quality
-
+    public static final int DEFAULT_MAX_SIDE_LENGTH_TRACK = 1008; // Reduce this to increase speed at the cost of mask quality
+    public static final int N_FRAME = -1; // number of frame to process
     // for image and video
     private double confidenceThreshold =DEFAULT_CONFIDENCE; // min probability (between 0 and 1) to add new detection
     private double maskScoreThreshold = DEFAULT_MASK_THRESHOLD; // min score (between -inf and +inf, centered on 0) to add pixel to binary mask
@@ -24,6 +24,7 @@ public class Sam3Parameters {
     private int removeAfterNMissed = DEFAULT_REMOVE_AFTER_MISSING; // if an object is not detected for x frames, it is removed from memory
     private int maxSideLengthDetect = DEFAULT_MAX_SIDE_LENGTH_DETECT;
     private int maxSideLengthTrack = DEFAULT_MAX_SIDE_LENGTH_TRACK;
+    private int nFrame = N_FRAME;
 
     // for image
     public Sam3Parameters() {
@@ -38,6 +39,7 @@ public class Sam3Parameters {
     public int getRemoveAfterNMissed() {return removeAfterNMissed;}
     public int getMaxSideLengthDetect() {return maxSideLengthDetect;}
     public int getMaxSideLengthTrack() {return maxSideLengthTrack;}
+    public int getNFrame() {return nFrame;}
 
     public void setConfidenceThreshold(double confidenceThreshold) {this.confidenceThreshold = confidenceThreshold;}
     public void setMaskScoreThreshold(double maskScoreThreshold) {this.maskScoreThreshold = maskScoreThreshold;}
@@ -48,4 +50,5 @@ public class Sam3Parameters {
     public void setRemoveAfterNMissed(int removeAfterNMissed) {this.removeAfterNMissed = removeAfterNMissed;}
     public void setMaxSideLengthDetect(int maxSideLengthDetect) {this.maxSideLengthDetect = maxSideLengthDetect;}
     public void setMaxSideLengthTrack(int maxSideLengthTrack) {this.maxSideLengthTrack = maxSideLengthTrack;}
+    public void setNFrame(int nFrame) {this.nFrame = nFrame;}
 }