+++ src/main/resources/fr/curie/miclearning/apposeplugin/sam3-detection-image-textprompt-m.py
@@ -102,7 +102,7 @@ if total_detection > 0:
     task.outputs['results_number'] = int(final_boxes.shape[0])
 
     # resize masks
-    final_masks = np.empty((total_detection, h_img, w_img), dtype=np.uint8, device="cpu")
+    final_masks = np.empty((total_detection, h_img, w_img), dtype=np.uint8)
     for i in range(total_detection):
         final_masks[i] = cv2.resize(pre_final_masks[i],(w_img, h_img),
                                       interpolation=cv2.INTER_NEAREST)